make Avatar component for reusable
